requirements.txt
duckdb>=0.9.0
pandas>=2.0.0
streamlit>=1.30.0
folium>=0.14.0
streamlit-folium>=0.11.0
tenacity>=8.2.0
httpx>=0.27.0
openpyxl>=3.1.0
APScheduler>=3.10.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
PyYAML>=6.0.0
pydeck>=0.9.0
anthropic>=0.95.0
fastapi>=0.100.0
uvicorn>=0.22.0
